compiler_rt: be more arch-specific
Only include architecture-specific files that we need. This is currently pretty
conservative in that e.g. not all div and mod functions are needed on all
architectures, but LLVM or subzero may want them so we leave them in for now.

This CL also adds support for multiply with overflow on ARM.

This CL depends on https://codereview.chromium.org/997813005/

R=jvoung@chromium.org, dschuff@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=4102
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3434
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3459

Review URL: https://codereview.chromium.org/997313002