Add __aeabi_idivmod to nacl-clang's compiler_rt
Add some asm helpers (using the nacl arm macros) to assembly.h, port
aeabi_idivmod.S and add it to the build.

Not every AEABI function will be able to be ported this way because some have
different calling conventions than normal functions. But this function
in particular has a user that needs it, so I started with it, and tested it
manually using nacl-gcc.

R=jfb@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=4140

Review URL: https://codereview.chromium.org/1054263002