Build ARM compiler-rt with EABI aliases
ARM EABI specifies a set of functions for the runtime support library, with
different names than those used by other platforms, as well as a few functions
not present elsewhere. Use of these has been disabled in PNaCl nacl-clang, but
it's better to be more conventional, especially for nacl-clang.
This CL switches nacl-clang's build to use the __aeabi_ aliases, and ports the
assembly implementations of some of the unique ones (e.g. __aeabi_ldivmod).
A few are left out, because they are not needed when running on hardware which
has VFP, which NaCl requires.

R=jvoung@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=4140

Review URL: https://codereview.chromium.org/1064113003